+++ data/embed.go
@@ -1,4 +1,4 @@
-package cache
+package data
 
 import _ "embed"
