+++ data/resources/genshin_characters.json
