+++ data/resources/genshin_profiles.json
