+++ data/resources/honkai_avatars.json
