+++ data/resources/honkai_characters.json
